requirements.txt
requests
openai
composio-openai
python-dotenv
httpx
